refactor(frontend): review follow-ups for the cutover
- The label and the provider mark always travel together and come from one
  resolution, so they resolve together: `engineIdentity()` returns both and
  `EnginePanelBody` takes the pair. A third piece of engine identity is now
  one edit rather than six.
- `EngineCardPrimitives` renamed to `EnginePanelPrimitives`. The engine card
  it was named after no longer exists; the file is the grid panels' tiles.
- The `logo` prop's doc read as the opposite of its meaning.
- The repo-layout tree omitted `components/engines/`, which survives the
  deletion as the engine panels' primitives and controls.
