+++ frontend/src/components/grid/panels/EngineSloGoodputPanel.tsx
@@ -1,10 +1,10 @@
-import { GoodputTile } from '@/components/engines/EngineCardPrimitives'
+import { GoodputTile } from '@/components/engines/EnginePanelPrimitives'
 import { SloSettingsControl } from '@/components/engines/SloSettingsControl'
 import { useSloSettings } from '@/hooks/useSloSettings'
 import { engineKey } from '@/lib/identity'
 import { combinedGoodput, formatSloThreshold, recomputeGoodputPct } from '@/lib/slo'
 import { EnginePanelBody } from './EnginePanelBody'
-import { engineLabel, engineLogo } from './engineLabel'
+import { engineIdentity } from './engineLabel'
 import { EnginePanelNotice } from './PanelNotice'
 import { useEnginePanel } from './useEnginePanel'
 import type { PanelContentProps } from '../panelRegistry'
@@ -47,8 +47,7 @@ export function EngineSloGoodputPanel({ panel }: PanelContentProps) {
 
   return (
     <EnginePanelBody
-      label={engineLabel(resolution)}
-      logo={engineLogo(resolution)}
+      identity={engineIdentity(resolution)}
       actions={
         <SloSettingsControl
           thresholds={thresholds}
